Refactoring ppmd8 decoder code
