+++ src/ext/Buffer.h
@@ -31,20 +31,8 @@ typedef struct {
     InBuffer *inBuffer;
 } BufferReader;
 
-typedef struct ppmd8_args_s {
-    CPpmd8 *cPpmd8;
-    InBuffer *in;
-    OutBuffer *out;
-    int max_length;
-    Bool finished;
-    int result;
-    pthread_t handle;
-} ppmd8_args;
 
 void Writer(const void *p, Byte b);
 Byte Reader(const void *p);
-
-Bool Ppmd8T_init(CPpmd8 *cPpmd8, UInt32 memory_size, unsigned int maximum_order, ISzAlloc *allocator);
-int Ppmd8T_decode(CPpmd8 *cPpmd8, OutBuffer *out, int max_length, ppmd8_args *args);
 
 #endif //PYPPMD_BUFFER_H
